Automagicalize conversion from numpy to eigen and dimension checking
We no longer expose `pybind11::array` types to the user. Instead. Input numpy types are transparently converted to Eigen with zero copy and the user accesses the variable as an Eigen type.

For example:

`npe_arg("a" "type_f32")` generates a variable `a` which is of type `Type_a` which is an `Eigen::Map`. The underlying `PlainObjectType` of `a` is accesible via the type `Matrix_a`.

We also do dimension checking propery and throw errors when the user passes in an invalid dimensional array.