Multiline docstrings.
There is sort of a shitty caveat. If you have a line in your docstring that starts with any amount of whitespace then `npe_*(`, you'll get a sort of cryptic error. I don't expect anybody to really do this so I'm going to let it slide and deal with it if ever there's an issue.